Move frontend in its own Django app
