+++ mocha/setup.js
@@ -36,7 +36,7 @@ global.WebSocket.prototype.close = function () {} // eslint-disable-line func-na
 global.Sinon = require("sinon").sandbox.create()
 
 // Trap calls to global library `Urls`
-const file = "socialhome/streams/app/tests/fixtures/Url"
+const file = "socialhome/frontend/tests/fixtures/Url"
 global.Urls = require(`../${file}`).Urls // eslint-disable-line global-require, import/no-dynamic-require
 
 global.gettext = key => key
