+++ socialhome/frontend/components/NsfwShield.vue
