+++ socialhome/frontend/components/stamped_elements/BaseStampedElement.vue
