+++ socialhome/frontend/components/stamped_elements/TagStampedElement.vue
@@ -5,7 +5,7 @@
 <script>
 import Vue from "vue"
 
-import "streams/app/components/stamped_elements/BaseStampedElement.vue"
+import "frontend/components/stamped_elements/BaseStampedElement.vue"
 
 
 export default Vue.component("tag-stamped-element", {
