+++ socialhome/frontend/stores/streamStore.js
@@ -6,8 +6,8 @@ import _defaults from "lodash/defaults"
 import _get from "lodash/get"
 import _pullAll from "lodash/pullAll"
 
-import getState from "streams/app/stores/streamStore.state"
-import {actions, mutations, streamStoreOperations, getters} from "streams/app/stores/streamStore.operations"
+import getState from "frontend/stores/streamStore.state"
+import {actions, mutations, streamStoreOperations, getters} from "frontend/stores/streamStore.operations"
 
 
 Vue.use(Vuex)
