+++ socialhome/frontend/tests/components/Stream.tests.js
@@ -4,12 +4,12 @@ import Vue from "vue"
 import BootstrapVue from "bootstrap-vue"
 import VueMasonryPlugin from "vue-masonry"
 
-import Stream from "streams/app/components/Stream.vue"
-import PublicStampedElement from "streams/app/components/stamped_elements/PublicStampedElement.vue"
-import FollowedStampedElement from "streams/app/components/stamped_elements/FollowedStampedElement.vue"
-import {streamStoreOperations} from "streams/app/stores/streamStore"
-import {getStore} from "streams/app/tests/fixtures/store.fixtures"
-import {getFakeContent} from "streams/app/tests/fixtures/jsonContext.fixtures"
+import Stream from "frontend/components/Stream.vue"
+import PublicStampedElement from "frontend/components/stamped_elements/PublicStampedElement.vue"
+import FollowedStampedElement from "frontend/components/stamped_elements/FollowedStampedElement.vue"
+import {streamStoreOperations} from "frontend/stores/streamStore"
+import {getStore} from "frontend/tests/fixtures/store.fixtures"
+import {getFakeContent} from "frontend/tests/fixtures/jsonContext.fixtures"
 import applicationStore from "../../stores/applicationStore"
 import {newStreamStore} from "../../stores/streamStore"
