+++ socialhome/frontend/tests/stores/streamStore.tests.js
@@ -4,10 +4,10 @@ import Moxios from "moxios"
 import Vue from "vue"
 import Vuex from "vuex"
 
-import {actions, mutations, getters} from "streams/app/stores/streamStore.operations"
-import {getFakeContent} from "streams/app/tests/fixtures/jsonContext.fixtures"
-import {streamStoreOperations, exportsForTests} from "streams/app/stores/streamStore"
-import getState from "streams/app/stores/streamStore.state"
+import {actions, mutations, getters} from "frontend/stores/streamStore.operations"
+import {getFakeContent} from "frontend/tests/fixtures/jsonContext.fixtures"
+import {streamStoreOperations, exportsForTests} from "frontend/stores/streamStore"
+import getState from "frontend/stores/streamStore.state"
 import Axios from "axios/index"
 import VueSnotify from "vue-snotify"
