+++ socialhome/streams/templates/streams/base.html
@@ -25,14 +25,14 @@
     {% json_context %}
     {{ block.super }}
     <script type="text/javascript" charset="utf-8">{% js_reverse_inline %}</script>
-    <script src="{% static 'js/webpack.stream.js' %}"></script>
+    <script src="{% static 'js/webpack.bundle.js' %}"></script>
     {% include "streams/includes/_twitter_widget.html" %}
 {% endblock %}
 
 
 {% block css %}
     {{ block.super }}
-    <link href="{% static "css/webpack.stream.css" %}" rel="stylesheet" >
+    <link href="{% static "css/webpack.bundle.css" %}" rel="stylesheet" >
     {# TODO: remove this override once legacy streams are removed #}
     {# Also change Content render to not add .nsfw class at all. #}
     <style>
